refactor : playbar 와 모달을 감싸는 컴포넌트 추가
