+++ src/domains/main/components/TopChartList.tsx
@@ -1,24 +1,12 @@
 "use client";
 import { useEffect, useState } from "react";
 
-import { Playlist, PlaylistItem } from "@/domains/playlist/types/Playlist";
-import PlayListModal from "@/domains/playlist/components/PlayListModal";
-import AddNewPlayListModal from "@/domains/playlist/components/AddNewPlayListModal";
-import PlaylistBar from "./PlaylistBar";
+import { PlaylistItem } from "@/domains/playlist/types/Playlist";
 
-export default function TopChartList({ isLoggedIn }: { isLoggedIn: boolean }) {
+import PlayBar from "./PlayBar";
+
+export default function TopChartList() {
   const [datas, setDatas] = useState<PlaylistItem[]>([]);
-  const [showPlayListModal, setShowPlayListModal] = useState(false);
-  const [showAddNewPlayListModal, setShowAddNewPlayListModal] = useState(false);
-  const [selectTrack, setSelectTrack] = useState<string[]>([]);
-  const [playlists, setPlaylists] = useState<Playlist[]>([]);
-
-  const handleShowPlayList = () => {
-    setShowPlayListModal(true);
-  };
-  const handleShowNewPlayList = () => {
-    setShowAddNewPlayListModal(true);
-  };
 
   useEffect(() => {
     fetch("/api/topChart")
@@ -29,33 +17,7 @@ export default function TopChartList({ isLoggedIn }: { isLoggedIn: boolean }) {
   return (
     <main>
       <h1>추천 음악을 들어보세요.</h1>
-      {datas.map((item) => (
-        <PlaylistBar
-          key={item.track.id}
-          item={item}
-          setSelectTrack={setSelectTrack}
-          handleShowPlayList={handleShowPlayList}
-        />
-      ))}
-      {isLoggedIn && (
-        <>
-          <PlayListModal
-            showModal={showPlayListModal}
-            onClose={() => setShowPlayListModal(false)}
-            playlists={playlists}
-            setPlaylists={setPlaylists}
-            onShowNewPlaylist={() => handleShowNewPlayList()}
-            track={selectTrack}
-          />
-
-          <AddNewPlayListModal
-            showModal={showAddNewPlayListModal}
-            onClose={() => setShowAddNewPlayListModal(false)}
-            setPlaylists={setPlaylists}
-            track={selectTrack}
-          />
-        </>
-      )}
+      <PlayBar tracks={datas} />
     </main>
   );
 }
